+++ Frontend/src/App.tsx
@@ -1,18 +1,28 @@
 import React from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
+import AuthenticatedLayout from "./components/layouts/AuthenticatedLayout";
+import GeneralLayout from "./components/layouts/GeneralLayout";
+import ErrorPage from "./components/ErrorPage";
 import Dashboard from "./components/dashboard/Dashboard";
 import Location from "./components/master/Location";
+import Login from "./components/Login";
+
 
 const App: React.FC = () => {
   return (
     <Router>
       <Routes>
-        {/* Dashboard Layout */}
-        <Route path="/" element={<Dashboard />}>
-          {/* Nested Routes */}
-          <Route path="dashboard" element={<div>Welcome to Dashboard!</div>} />
-          <Route path="settings/location" element={<Location />} />
+        {/* General Route */}
+        <Route element={<GeneralLayout />}>
+          <Route path="/login" element={<Login />} />
         </Route>
+        {/* Authenticated Route */}
+        <Route element={<AuthenticatedLayout />}>
+          <Route path="/dashboard" element={<Dashboard />} />
+          <Route path="/settings/location" element={<Location />} />
+        </Route>        
+        {/* Error Route */}
+        <Route path="*" element={<ErrorPage />} />
       </Routes>
     </Router>
   );
