+++ Frontend/src/components/ErrorPage.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+const ErrorPage: React.FC = () => {
+  return(
+    <div style={{textAlign: "center", padding: "50px"}}>
+      <h1>404 - Page Not Found</h1>
+      <p>The page you are looking for does not exist.</p>
+    </div>
+  );
+};
+
+export default ErrorPage;
