+++ Frontend/src/components/layouts/GeneralLayout.css
@@ -1,0 +1,7 @@
+.layout {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100vh;
+  background-color: #eaeaea;
+}
