+++ Frontend/src/components/layouts/GeneralLayout.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { Outlet } from "react-router-dom";
+
+const GeneralLayout: React.FC = () => {
+  return(
+    <div className="layout">
+      <Outlet />
+    </div>
+  );
+};
+
+export default GeneralLayout;
